pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "alpha-vantage>=3.0.0",
    "fastapi>=0.115.11",
    "jwt>=1.3.1",
    "numpy>=2.2.3",
    "pandas>=2.2.3",
    "passlib[bcrypt]>=1.7.4",
    "plotly>=6.0.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.1",
    "python-jose[cryptography]>=3.4.0",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.39",
    "streamlit>=1.43.2",
    "uvicorn>=0.34.0",
]
